+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 NA
 
+## 0.2.0 (2021-10-30)
+
+* Use `ptime` for time rather than `Core`
+* `Level.of_string` returns `t option` now rather than `t Or_error.t`
+* Drop `ppx_compare`
+
 ## 0.1.1 (2021-08-18)
 
 Fixed some dependency problems in the `opam` file.